+++ C-Plus-Plus/README.md
@@ -442,7 +442,7 @@ _add list here_
 
 ## Other
 
-- [Transpose of a 3X3 Matrix](other/Transpose_Of_3X3_Matrix.cpp)
+- [Transpose of a Square Matrix](other/Transpose_Of_Square_Matrix.cpp)
 - [Add Two Linked Lists](other/addTwoLinkedLists.cpp)
 - [Arranging elements of array to form biggest possible number](other/Arranging_elements_of_array_to_form_biggestno.cpp)
 - [Boyer–Moore majority vote algorithm](other/majority_vote_algorithm.cpp)
